/authors end point work done

+++ src/main/java/com/rehan/librarymanagementsystem/service/AuthorService.java
@@ -1,8 +1,11 @@
 package com.rehan.librarymanagementsystem.service;
 
+import com.rehan.librarymanagementsystem.exceptions.AuthorNotFoundException;
 import com.rehan.librarymanagementsystem.model.Author;
 import com.rehan.librarymanagementsystem.repositories.AuthorRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.Optional;
 
 @Service
 public class AuthorService {
@@ -18,14 +21,25 @@ public class AuthorService {
     }
 
     public Author findById(int id) {
-        return authorRepository.findById(id).get();
+        Optional<Author> author = authorRepository.findById(id);
+        if(author.isEmpty()) throw new AuthorNotFoundException("Author with authorId : " + id  +" does not exists");
+        return author.get();
     }
 
     public void deleteById(int id) {
+        Optional<Author> author = authorRepository.findById(id);
+        if(author.isEmpty()) throw new AuthorNotFoundException("Author with authorId : " + id  +" does not exists");
         authorRepository.deleteById(id);
     }
 
-    public Author save(Author author) {
+    public Author addNewAuthor(Author author) {
+        author.setAuthorId(0);
+        return authorRepository.save(author);
+    }
+
+    public Author updateAuthor(Author author) {
+        int authorId = author.getAuthorId();
+        if(authorRepository.findById(authorId).isEmpty()) throw new AuthorNotFoundException("Author with authorId : "+authorId+" does not exists");
         return authorRepository.save(author);
     }
 }
